IborIndex class

This can be used to represent IBOR like indices (e.g. LIBOR, BBSW, CDOR) and extends the Index class.

IborIndex(name, currency, tenor, spot_lag, calendar, day_basis,
  day_convention, is_eom)

Arguments

name
the name of the index as a string

currency
the currency associated with the index as a Currency object

tenor
the term of the index as a period

spot_lag
the period between the index's fixing and the start of the index's term

calendar
the calendar used to determine whether the index fixes on a given date as a Calendar

day_basis
the day basis associated with the index (e.g. "act/365")

day_convention
the day convention associated with the index (e.g. "mf")

is_eom
a flag indicating whether or not the maturity date of the index is subject to the end-to-end convention.

Value

an object of class IborIndex that inherits from Index

Examples

library(lubridate)
library(fmdates)
# 3m AUD BBSW
IborIndex("BBSW", AUD(), months(3), days(0), c(AUSYCalendar()),
  "act/365", "ms", FALSE)

#> <IborIndex> 3m AUD BBSW