SingleCurrencyMoney

This class associates a numeric vector with a currency. This is useful for example in representing the value of a derivative. You can concatenate a set SingleCurrencyMoney objects and return a MultiCurrencyMoney object (see example below)

SingleCurrencyMoney(value, currency)

Arguments

value
a numeric vector of values

currency
a single Currency object

Value

a SingleCurrencyMoney object

See also

Examples

SingleCurrencyMoney(1:5, AUD())

#> <SingleCurrencyMoney> AUD 1 2 3 4 5 

c(SingleCurrencyMoney(1, AUD()), SingleCurrencyMoney(100, USD()))

#> # <MultiCurrencyMoney> of length 2
#>   values currencies
#> *  <dbl> <list>    
#> 1      1 <CCY: AUD>
#> 2    100 <CCY: USD>