VolSurface class

The VolSurface class is designed to capture implied volatility information along with information about how to interpolate an implied volatility between nodes.

VolSurface(vol_quotes, interpolation)

Arguments

vol_quotes
object of class VolQuotes() containing the volatility data.

interpolation
Interplation method, given as an object of class interpolation Interpolation(). At this time only LinearCubicTimeVarInterpolation() is supported. This is a two-dimensional interpolator that uses linear interpolation in the time dimension and cubic splines in the smile dimension with the values interpolated being the square of the implied volatilities. Return values are implied volatilies

Value

a VolSurface object

See also

Examples

VolSurface(build_vol_quotes(), LinearCubicTimeVarInterpolation())

#> <VolSurface> @ 26 April 2019 
#>  ABC.AX   
#> # A tibble: 867 x 3
#>    maturity   smile value
#>    <date>     <dbl> <dbl>
#>  1 2019-04-27  4.88 0.798
#>  2 2019-05-30  4.88 0.672
#>  3 2019-06-27  4.88 0.546
#>  4 2019-09-26  4.88 0.674
#>  5 2019-12-19  4.88 0.638
#>  6 2020-06-30  4.88 0.581
#>  7 2020-12-31  4.88 0.588
#>  8 2021-06-30  4.88 0.617
#>  9 2021-12-31  4.88 0.620
#> 10 2022-06-30  4.88 0.617
#> # … with 857 more rows