ZeroCurve class

A class that defines the bare bones of a zero-coupon yield curve pricing structure.

ZeroCurve(discount_factors, reference_date, interpolation)

Arguments

discount_factors
a DiscountFactor object. These are converted to continuously compounded zero coupon interest rates with an act/365 day basis for internal storage purposes

reference_date
a Date object

interpolation
an Interpolation object

Value

a ZeroCurve object

Details

A term structure of interest rates (or yield curve) is a curve showing several yields or interest rates across different contract lengths (2 month, 2 year, 20 year, etc...) for a similar debt contract. The curve shows the relation between the (level of) interest rate (or cost of borrowing) and the time to maturity, known as the "term", of the debt for a given borrower in a given currency. For example, the U.S. dollar interest rates paid on U.S. Treasury securities for various maturities are closely watched by many traders, and are commonly plotted on a graph. More formal mathematical descriptions of this relation are often called the term structure of interest rates. When the effect of coupons on yields are stripped away, one has a zero-coupon yield curve.

The following interpolation schemes are supported by ZeroCurve: ConstantInterpolation, LinearInterpolation, LogDFInterpolation and CubicInterpolation. Points outside the calibration region use constant extrapolation on the zero rate.

See also

Interpolation

Examples

zc_df <- fmdata_example("zerocurve.csv")

#> Error in fmdata_example("zerocurve.csv"): could not find function "fmdata_example"

values <- zc_df[["dfs"]]

#> Error in eval(expr, envir, enclos): object 'zc_df' not found

starts <- as.Date(as.character(zc_df[["start"]]), "%Y%m%d")

#> Error in as.Date(as.character(zc_df[["start"]]), "%Y%m%d"): object 'zc_df' not found

ends <- as.Date(as.character(zc_df[["end"]]), "%Y%m%d")

#> Error in as.Date(as.character(zc_df[["end"]]), "%Y%m%d"): object 'zc_df' not found

dfs <- DiscountFactor(values, starts, ends)

#> Error in NROW(value): object 'values' not found

zc <- ZeroCurve(dfs, starts[1], LogDFInterpolation())

#> Error in lubridate::as_date(date1): object 'starts' not found

plot(zc$pillar_times, zc$pillar_zeros, xlab = 'Years', ylab = 'Zero')

#> Error in plot(zc$pillar_times, zc$pillar_zeros, xlab = "Years", ylab = "Zero"): object 'zc' not found