ZeroCurve attributes as a data frame

Create a tibble that contains the pillar point maturities in years (using the act/365 convention) and the corresponding continuously compounded zero rates.

# S3 method for ZeroCurve
as_tibble(x, ...)

Arguments

x
a ZeroCurve object

...
other parameters that are not used by this methods

Value

a tibble with two columns named Years and Zeros.

See also

tibble::tibble()

Examples

library(tibble)
zc <- build_zero_curve()
as_tibble(zc)

#> # A tibble: 27 x 4
#>       start      end zeros   dfs
#>       <int>    <int> <dbl> <dbl>
#>  1 20151231 20160104  2.01 1.000
#>  2 20160104 20160112  2.01 0.999
#>  3 20160104 20160119  2.01 0.999
#>  4 20160104 20160205  2.01 0.998
#>  5 20160104 20160307  1.99 0.996
#>  6 20160104 20160405  1.98 0.995
#>  7 20160104 20160505  1.96 0.993
#>  8 20160104 20160607  1.95 0.992
#>  9 20160104 20160705  1.93 0.990
#> 10 20160104 20161005  1.91 0.986
#> # ... with 17 more rows