Source supplied financial market data

The following files are supplied by the package:

zerocurve.csv has four fields for one curve: start, end, zeros and dfs representing the start and end dates of the pillar instruments and semi-annually compounded zero coupon rates and discount factors

zerocurves.csv has five fields for curve sets associated with two currencies: start, end, zeros and dfs as above as well as name representing the name of the curve. The name of the curve is specified as CCY_INDEX where CCY is the ISO code of the curve's currency and INDEX is the index associated with the curve.

fmdata_example(file)

Arguments

file
the name of the file containing the data set.

Value

a tibble

See also

Other build object helpers: build_zero_curve

Examples

fmdata_example("zerocurve.csv")

#> # A tibble: 27 x 4
#>       start      end zeros   dfs
#>       <int>    <int> <dbl> <dbl>
#>  1 20151231 20160104  2.01 1.000
#>  2 20160104 20160112  2.01 0.999
#>  3 20160104 20160119  2.01 0.999
#>  4 20160104 20160205  2.01 0.998
#>  5 20160104 20160307  1.99 0.996
#>  6 20160104 20160405  1.98 0.995
#>  7 20160104 20160505  1.96 0.993
#>  8 20160104 20160607  1.95 0.992
#>  9 20160104 20160705  1.93 0.990
#> 10 20160104 20161005  1.91 0.986
#> # ... with 17 more rows