Standard IBOR

These functions create commonly used IBOR indices with standard market conventions.

AUDBBSW(tenor)

AUDBBSW1b(tenor)

EURIBOR(tenor)

GBPLIBOR(tenor)

JPYLIBOR(tenor)

JPYTIBOR(tenor)

NZDBKBM(tenor)

USDLIBOR(tenor)

CHFLIBOR(tenor)

HKDHIBOR(tenor)

NOKNIBOR(tenor)

Arguments

tenor
the tenor of the IBOR index (e.g. months(3))

Details

The key conventions are tabulated below.

Creator 	 Spot lag (days) 	 Fixing calendars 	 Day basis 	 Day convention 	 EOM

AUDBBSW() 	 0 	 AUSYCalendar 	 act/365 	 ms 	 FALSE

EURIBOR() 	 2 	 EUTACalendar 	 act/360 	 mf 	 TRUE

GBPLIBOR() 	 0 	 GBLOCalendar 	 act/365 	 mf 	 TRUE

JPYLIBOR() 	 2 	 GBLOCalendar 	 act/360 	 mf 	 TRUE

JPYTIBOR() 	 2 	 JPTOCalendar 	 act/365 	 mf 	 FALSE

NZDBKBM() 	 0 	 NZWECalendar, NZAUCalendar 	 act/365 	 mf 	 FALSE

USDLIBOR() 	 2 	 USNYCalendar, GBLOCalendar 	 act/360 	 mf 	 TRUE

CHFLIBOR() 	 2 	 GBLOCalendar 	 act/360 	 mf 	 TRUE

HKDHIBOR() 	 0 	 HKHKCalendar 	 act/365 	 mf 	 FALSE

There are some nuances to this. Sub-1m LIBOR and TIBOR spot lags are zero days (excepting spot-next rates) and use the following day convention and the overnight USDLIBOR index uses both USNYCalendar and GBLOCalendar calendars.

References

BBSW EURIBOR ICE LIBOR BBA LIBOR TIBOR NZD BKBM OpenGamma Interest Rate Instruments and Market Conventions Guide HKD HIBOR

See also

Other constructors: CurrencyConstructors, CurrencyPairConstructors, oniaindices