Interpolate zeros

This interpolates zero rates from either a ZeroCurve or some other object that contains such an object.

interpolate_zeros(x, at, compounding = NULL, day_basis = NULL, ...)

# S3 method for ZeroCurve
interpolate_zeros(x, at, compounding = NULL,
  day_basis = NULL, ...)

Arguments

x
the object to interpolate

at
a Date vector representing the date at which to interpolate a value

compounding
a valid compounding string. Defaults to NULL which uses the curve's native compounding basis

day_basis
a valid day basis string. Defaults to NULL which uses the curve's native day basis.

...
further arguments passed to specific methods

Value

an InterestRate object of interpolated zero rates with the compounnding and day_basis requested.

See also