ZeroCurve class

A class that defines the bare bones of a zero-coupon yield curve pricing structure.

ZeroCurve(discount_factors, reference_date, interpolation)

Arguments

discount_factors
a DiscountFactor object. These are converted to continuously compounded zero coupon interest rates with an act/365 day basis for internal storage purposes

reference_date
a Date object

interpolation
an Interpolation object

Value

a ZeroCurve object

Details

A term structure of interest rates (or yield curve) is a curve showing several yields or interest rates across different contract lengths (2 month, 2 year, 20 year, etc...) for a similar debt contract. The curve shows the relation between the (level of) interest rate (or cost of borrowing) and the time to maturity, known as the "term", of the debt for a given borrower in a given currency. For example, the U.S. dollar interest rates paid on U.S. Treasury securities for various maturities are closely watched by many traders, and are commonly plotted on a graph. More formal mathematical descriptions of this relation are often called the term structure of interest rates. When the effect of coupons on yields are stripped away, one has a zero-coupon yield curve.

Interpolation schemes

The following interpolation schemes are supported by ZeroCurve:

ConstantInterpolation: constant interpolation on zero rates

LinearInterpolation: linear interpolation on zero rates

LogDFInterpolation: linear interpolation on log discount factors or constant on forward rates

CubicInterpolation: natural cubic spline on zero rates

Points outside the calibration region use constant extrapolation on zero rates.

See also

Examples

build_zero_curve()

#> <ZeroCurve> @ 31 December 2015 
#> # A tibble: 27 x 2
#>     Years  Zeros
#>     <dbl>  <dbl>
#>  1 0.0110 0.0200
#>  2 0.0329 0.0200
#>  3 0.0521 0.0200
#>  4 0.0986 0.0200
#>  5 0.184  0.0198
#>  6 0.263  0.0197
#>  7 0.345  0.0195
#>  8 0.436  0.0194
#>  9 0.512  0.0192
#> 10 0.764  0.0190
#> # … with 17 more rows